backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.14",
    "clerk-backend-api>=3.0.5",
    "langchain>=0.3.26",
    "langchain-ollama>=0.3.4",
    "langchain-openai>=0.3.27",
    "langfuse>=3.1.3",
    "neo4j>=5.28.1",
    "ollama>=0.5.1",
    "pydantic>=2.11.7",
    "sentence-transformers>=5.0.0",
    "tenacity>=9.1.2",
    "python-frontmatter>=1.1.0",
    "tiktoken>=0.8.0",
    "redis>=6.2.0",
    "rq>=2.4.1",
    "langchain-google-genai>=2.1.8",
]

[dependency-groups]
dev = ["black>=24.0.0"]

[tool.black]
line-length = 88
target-version = ['py313']
